+++ packages/verified-fetch/src/utils/server-timing.ts
@@ -1,10 +1,8 @@
 export class ServerTiming {
   private headers: string[]
-  private precision: number
 
   constructor () {
     this.headers = []
-    this.precision = 3
   }
 
   getHeader (): string {
@@ -18,13 +16,13 @@ export class ServerTiming {
       return await promise // Execute the function
     } finally {
       const endTime = performance.now()
-      const duration = (endTime - startTime).toFixed(this.precision) // Duration in milliseconds
+      const duration = endTime - startTime
 
       this.add(name, description, duration)
     }
   }
 
-  add (name: string, description: string, duration: number | string): void {
-    this.headers.push(`${name};dur=${Number(duration).toFixed(this.precision)}${description === '' ? '' : `;desc="${description}"`}`)
+  add (name: string, description: string, ms: number): void {
+    this.headers.push(`${name};dur=${Math.round(ms)}${description === '' ? '' : `;desc="${description}"`}`)
   }
 }